requirements.txt
pandas
numpy
seaborn
matplotlib
sklearn.preprocessing
catboost
xgboost
dataclasses
scikit-learn==0.24.2
#-e .